feat: make alt page the new home page and move current home page to alt and set standard header size
