+++ app/components/Home/ValuesSection.vue
