+++ app/components/TheHeader.vue
@@ -25,7 +25,6 @@ const activeJourney = computed(() => route.path.startsWith("/journey"));
 <template>
   <cv-header
     id="ff-header"
-    class="!h-24"
   >
     <cv-header-menu-button
       id="ff-header-menu"
@@ -40,7 +39,7 @@ const activeJourney = computed(() => route.path.startsWith("/journey"));
       <img
         alt=""
         src="@/assets/images/FiveFifthsVoterLogo.png"
-        class="h-full max-h-[32px] sm:max-h-[64px]"
+        class="h-full"
       >
       FiveFifths
     </cv-header-name>
@@ -81,7 +80,7 @@ const activeJourney = computed(() => route.path.startsWith("/journey"));
     </cv-header-nav>
     <template #header-global>
       <cv-link
-        class="px-2 py-8"
+        class="p-2"
         href="https://www.instagram.com/fivefifthsvoter/?igshid=Zjc2ZTc4Nzk%3D"
         target="_blank"
         aria-label="Five Fifths Voter on instagram - link opens in a new window"
@@ -89,7 +88,7 @@ const activeJourney = computed(() => route.path.startsWith("/journey"));
         <Instagram class="text-carbon-gray-30" />
       </cv-link>
       <cv-link
-        class="px-2 py-8"
+        class="p-2"
         href="https://github.com/Call-for-Code-for-Racial-Justice/Five-Fifths-Voter"
         target="_blank"
         aria-label="Link to code for this site in GitHub - link opens in a new window"
@@ -101,7 +100,6 @@ const activeJourney = computed(() => route.path.startsWith("/journey"));
         :aria-label="$t('ariaLanguageSetting')"
         aria-controls="language-panel"
         :label="$t('ariaLanguageSetting')"
-        class="!mx-2 !h-6 !min-h-0 !px-0 !py-12"
         tip-position="left"
       >
         <LanguageSwitcher class="text-carbon-gray-30" />
